requirements.txt
absl-py==0.9.0
appdirs==1.4.4
Babel==2.8.0
certifi==2020.6.20
cffi==1.11.5
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.2.1
cycler==0.10.0
decorator==4.4.2
distlib==0.3.0
docopt==0.6.2
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Cors==3.0.8
funcsigs==1.0.2
future==0.18.2
graphviz==0.14
grpcio==1.30.0
gym==0.17.2
hdfs==2.5.8
identify==1.4.20
idna==2.9
importlib-metadata==1.6.1
importlib-resources==2.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
nltk==3.5
nodeenv==1.4.0
numpy==1.19.0
objgraph==3.4.1
opencv-python==4.2.0.34
paddlepaddle-gpu==1.6.3.post107
parl==1.3.1
Pillow==7.1.2
-e git+https://github.com/ntasfi/PyGame-Learning-Environment.git@3dbe79dc0c35559bb441b9359948aabf9bb3d331#egg=ple
pre-commit==2.5.1
prettytable==0.7.2
protobuf==3.12.2
psutil==5.7.0
pyarrow==0.13.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pygame==1.9.6
pyglet==1.5.0
pylocker==3.0.0
pyparsing==2.4.7
pyquaternion==0.9.5
PyRep==1.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==18.0.1
rarfile==3.1
regex==2020.6.8
requests==2.24.0
rlbench==1.0.7
scipy==1.2.1
six==1.15.0
tb-nightly==1.15.0a20190801
tensorboardX==1.8
termcolor==1.1.0
toml==0.10.1
tqdm==4.46.1
urllib3==1.25.9
virtualenv==20.0.25
visualdl==2.0.0b7
Werkzeug==1.0.1
zipp==3.1.0
